update some files

+++ shell/mysql_install.sh
@@ -1,13 +1,16 @@
 #!/bin/bash
-
+# 2022.10.1更新,优化多实例
+# # 修改basedir路径,使得不同实例可以使用相同的basedir,节省磁盘空间
+# # 修改my.cnf路径,方便多实例管理
 # 适用于mysqld 5.7/8.0
 # 修改以下变量
 download_mysql_version=8.0.28 # 要安装的版本
-mysql_listen_port=3306 # 此实例监听的端口
+mysql_listen_port=3306 # 此实例监听的端口(影响socket、数据目录、配置文件路径、服务名)
 mysql_data_dir=/mysqld/data_${mysql_listen_port} # 数据目录
-mysql_base_dir=/usr/local/mysql_${mysql_listen_port} # 安装目录
+mysql_base_dir=/usr/local/mysql_${download_mysql_version} # 安装目录
+mysql_conf_dir=/mysqld/etc_${mysql_listen_port} # 配置文件路径
 serverid=10 # 集群id
-root_passwd=123456 # 安装完成后将root密码修改为此字符
+root_passwd=123456 # 安装完成后将root密码修改为此字符串
 
 init_environment() {
     data_dir=$1
@@ -27,8 +30,8 @@ init_environment() {
             rm -rf /etc/my.cnf /etc/mysql && echo "/etc/my.cnf or /etc/mysql already deleted"
         fi
     fi
-    if [ -e ${base_dir} ] || [ -e ${data_dir}/mysql ];then
-        echo "mysql already exists. please check ${base_dir} and ${data_dir}" && exit
+    if [ -e ${data_dir}/mysql ];then
+        echo "mysql already exists. please check ${data_dir}" && exit
     fi
     id mysql >/dev/null 2>&1
     if [ $? -ne 0 ];then
@@ -67,20 +70,26 @@ install_mysql() {
     mysql_port=$4
     server_id=$5
     os_version=$6
-    if [ -e /data/work/mysql-${version}-linux-glibc2.12-x86_64.tar.xz ] && [[ ${version:0:3} = 8.0 ]];then
-        tar -xf /data/work/mysql-${version}-linux-glibc2.12-x86_64.tar.xz -C /data/work/
-    elif [ -e /data/work/mysql-${version}-linux-glibc2.12-x86_64.tar.gz ];then
-        tar -zxf /data/work/mysql-${version}-linux-glibc2.12-x86_64.tar.gz -C /data/work/
+    mysql_conf_dir=$7
+    if [ ! -e ${base_dir} ];then
+        if [ -e /data/work/mysql-${version}-linux-glibc2.12-x86_64.tar.xz ] && [[ ${version:0:3} = 8.0 ]];then
+            tar -xf /data/work/mysql-${version}-linux-glibc2.12-x86_64.tar.xz -C /data/work/
+        elif [ -e /data/work/mysql-${version}-linux-glibc2.12-x86_64.tar.gz ];then
+            tar -zxf /data/work/mysql-${version}-linux-glibc2.12-x86_64.tar.gz -C /data/work/
+        else
+            download_mysql ${version}
+            echo "Downloading mysql-${version}-glibc-install package,when download is completed,please run this script again."
+            exit
+        fi
+        mv /data/work/mysql-${version}-linux-glibc2.12-x86_64 ${base_dir}
+        mkdir -p ${base_dir}/mysql-files
+        chown -R mysql:mysql ${base_dir}
+        chmod 750 ${base_dir}/mysql-files
     else
-        download_mysql ${version}
-        echo "Downloading mysql-${version}-glibc-install package,when download is completed,please run this script again."
-        exit
+        echo "${base_dir} already exists"
     fi
-    mv /data/work/mysql-${version}-linux-glibc2.12-x86_64 ${base_dir}
-    mkdir -p ${base_dir}/mysql-files
-    chown mysql:mysql ${base_dir}/mysql-files
-    chmod 750 ${base_dir}/mysql-files
-    cat > ${base_dir}/my.cnf <<EOF
+    mkdir -p ${mysql_conf_dir}
+    cat > ${mysql_conf_dir}/my.cnf <<EOF
 [mysqld]
 # base
 basedir=${base_dir}
@@ -151,7 +160,7 @@ replica_parallel_type=LOGICAL_CLOCK
 replica_parallel_workers=8
 log_replica_updates=1
 EOF
-    chown -R mysql:mysql ${base_dir}/my.cnf
+    chown -R mysql:mysql ${mysql_conf_dir}
     if [ ! -e /etc/my.cnf ] && [ ! -e /etc/mysql ];then
         cat > /etc/my.cnf <<EOF
 [mysql]
@@ -167,7 +176,7 @@ EOF
     else
         echo "/etc/my.cnf or /etc/mysql already exists,please check them."
     fi
-    ${base_dir}/bin/mysqld --defaults-file=${base_dir}/my.cnf --initialize --user=mysql --basedir=${base_dir} --datadir=${data_dir} 2>&1 | tee -a /tmp/my_install_${mysql_port}.log
+    ${base_dir}/bin/mysqld --defaults-file=${mysql_conf_dir}/my.cnf --initialize --user=mysql 2>&1 | tee -a /tmp/my_install_${mysql_port}.log
     ${base_dir}/bin/mysql_ssl_rsa_setup --datadir=${data_dir}
     if [ -e ${data_dir}/undo_001 ];then
         mkdir ${data_dir}/undospace/ -p
@@ -179,6 +188,7 @@ EOF
     chown -R mysql:mysql ${data_dir}
     cp ${base_dir}/support-files/mysql.server /etc/init.d/mysqld_${mysql_port}
     sed -i -e "/^datadir=/s#datadir=#datadir=${data_dir}#g" \
+    -e "/^other_args=/s#other_args=\"\$\*\"#other_args=\"\`\$basedir/bin/my_print_defaults --defaults-extra-file=${mysql_conf_dir}/my.cnf mysqld\` \$\*\"#g" \
     -e "/^basedir=/s#basedir=#basedir=${base_dir}#g" /etc/init.d/mysqld_${mysql_port}
     chmod +x /etc/init.d/mysqld_${mysql_port}
     systemctl daemon-reload
@@ -255,6 +265,7 @@ check_configuration() {
     base_dir=$4
     mysql_port=$5
     server_id=$6
+    mysql_conf=$7
     echo -e "\
 please check configuration \
 \n  os_version: ${os_version} \
@@ -262,7 +273,8 @@ please check configuration \
 \n  mysqldatadir: ${data_dir} \
 \n  mysql_basedir: ${base_dir} \
 \n  listen_port: ${mysql_port} \
-\n  server_id: ${server_id}"
+\n  server_id: ${server_id} \
+\n  my.cnf: ${mysql_conf_dir}"
     read -p "Do you confirm these configuration(y/n): " action
     if [[ ${action} != y ]];then
         echo "[warning] please modify $0 configuration." && exit
@@ -291,7 +303,7 @@ do
         1)
             init_environment ${mysql_data_dir} ${mysql_base_dir} ${os_version} 
             download_mysql ${download_mysql_version}
-            install_mysql ${download_mysql_version} ${mysql_data_dir} ${mysql_base_dir} ${mysql_listen_port} ${serverid} ${os_version} 
+            install_mysql ${download_mysql_version} ${mysql_data_dir} ${mysql_base_dir} ${mysql_listen_port} ${serverid} ${os_version} ${mysql_conf_dir}
             change_random_passwd ${mysql_listen_port} ${root_passwd}
         ;;
         2)
@@ -301,7 +313,7 @@ do
             mysql_auto_start_disable ${mysql_listen_port} ${os_version} 
         ;;
         4)
-            check_configuration ${os_version} ${download_mysql_version} ${mysql_data_dir} ${mysql_base_dir} ${mysql_listen_port} ${serverid}
+            check_configuration ${os_version} ${download_mysql_version} ${mysql_data_dir} ${mysql_base_dir} ${mysql_listen_port} ${serverid} ${mysql_conf_dir}
         ;;
         q)
             break
